Make node in containers see each other
Problem: Currently nodes inside the containers use `locahost` as
net-addr and rpc-addr which makes it impossible for them to see each
other.

Solution: There is an easy way to get the ip address of the docker
container within the docker network (it communicates all containers and
the host). Use this container ip address in the scripts instead of
net-addr and rpc-addr.